requirements.txt
spacy==3.1.4
pdfminer.six==20201018